Создает компонент BoardController и переносит всю логику по отрисовке задач в него.
